FIX: 🪲️Refactor ApiExceptionListener into an event subscriber and reorganize its namespace
